+++ src/Microsoft.ML.Dnn/DnnCatalog.cs
@@ -6,10 +6,10 @@ using Microsoft.ML.Data;
 using Microsoft.ML.Dnn;
 using Microsoft.ML.Runtime;
 using Microsoft.ML.Transforms;
+using static Microsoft.ML.TensorFlow.TensorFlowUtils;
 
 namespace Microsoft.ML
 {
-    /// <include file='doc.xml' path='doc/members/member[@name="DnnRetrainTransformer"]/*' />
     public static class DnnCatalog
     {
 
@@ -75,7 +75,7 @@ namespace Microsoft.ML
             };
 
             var env = CatalogUtils.GetEnvironment(catalog);
-            return new DnnRetrainEstimator(env, options, DnnUtils.LoadDnnModel(env, modelPath, true));
+            return new DnnRetrainEstimator(env, options, LoadDnnModel(env, modelPath, true));
         }
 
         /// <summary>
